// generated/src/app/api-models/sec-win.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SecWin {
  userId: string;
  winId: string;
  pSel: string;
  pIns: string;
  pUpd: string;
  pDel: string;
  maxOpen: number;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  languageId: number;
}